requirements.txt
matplotlib
pandas
numpy
shapely
python_sat
pyyaml